practica5

+++ tipos de datos/practica5.py
@@ -1,3 +1,3 @@
 numeroh = int(input("Ingrese un número de hors que trabajas: "))
-costo_h = int(input("Ingrese el costo por hora: ")) 
+costo_h = float(input("Ingrese el costo por hora: ")) 
 print(f"El total a pagar es: {numeroh * costo_h}$")
